Create form to create user from admin_entreprise role, create function instanceOfEntreprise to show users from the same company only

+++ src/Controller/UsersEntreprisesController.php
@@ -2,7 +2,6 @@
 
 namespace App\Controller;
 
-use App\Form\UserType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -10,6 +9,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Entity\User;
+use App\Form\UserEntreprisesType;
 use App\Repository\UserRepository;
 
 #[Route('/users')]
@@ -22,9 +22,21 @@ class UsersEntreprisesController extends AbstractController
     {
         $this->passwordHasher = $passwordHasher;
     }
-    
 
-    #[Route('/', name: 'app_entreprises_index', methods: ['GET'])]
+    private function instanceOfEntreprise(User $user): bool
+    {
+        $currentUser = $this->getUser();
+
+        if (!$currentUser instanceof User) {
+            return false;
+        }
+
+        $entreprise = $currentUser->getIdEntreprise();
+
+        return $user->getIdEntreprise() === $entreprise;
+    }
+
+    #[Route('/', name: 'app_users_index', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
     {
         
@@ -57,7 +69,7 @@ class UsersEntreprisesController extends AbstractController
 
         $newUser = new User();
 
-        $form = $this->createForm(UserType::class, $newUser);
+        $form = $this->createForm(UserEntreprisesType::class, $newUser);
 
         $form->handleRequest($request);
 
@@ -82,6 +94,11 @@ class UsersEntreprisesController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        
+        if (!$this->instanceOfEntreprise($user)) {
+            return $this->redirectToRoute('app_users_index');
+        }
+
         return $this->render('usersEntreprises/show.html.twig', [
             'user' => $user,
         ]);
@@ -90,7 +107,11 @@ class UsersEntreprisesController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
-        $form = $this->createForm(UserType::class, $user);
+        if (!$this->instanceOfEntreprise($user)) {
+            return $this->redirectToRoute('app_users_index');
+        }
+
+        $form = $this->createForm(UserEntreprisesType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -101,7 +122,7 @@ class UsersEntreprisesController extends AbstractController
             return $this->redirectToRoute('app_account');
         }
 
-        return $this->render('users/edit.html.twig', [
+        return $this->render('usersEntreprises/edit.html.twig', [
             'user' => $user,
             'form' => $form,
         ]);
@@ -110,6 +131,10 @@ class UsersEntreprisesController extends AbstractController
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->instanceOfEntreprise($user)) {
+            return $this->redirectToRoute('app_users_index');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager->remove($user);
             $entityManager->flush();
